feat(SFT-458): renaming files for field types
